Updated allure screenshots

+++ generate_report.bat
@@ -1,3 +1,3 @@
-rmdir /s /q allure-results
+if exist allure-results rmdir /s /q allure-results
 pytest --alluredir allure-results
 allure serve allure-results
